backend/pyproject.toml
[project]
name = "brochure-api"
version = "0.1.0"
description = "FastAPI backend for Business Brochure Generator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "beautifulsoup4>=4.14.2",
    "requests>=2.32.5",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]
